% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abcRej.R
\name{abcRej}
\alias{abcRej}
\title{Approximate Bayesian Computation inference via Rejection Method}
\usage{
abcRej(x, sim.model, prior, nsim, tol, ncore = 1)
}
\arguments{
\item{x}{A cFreqData class object with the observed variant frequencies.}

\item{sim.model}{A transmission model generated using the \code{genSim()} function.}

\item{prior}{A list of model priors. Each element of the list corresponds to a model parameter. The list element must be a vector whose first argument determines the type of prior distribution: possible values are "unif" for a uniform distribution on a segment and "normal" for a normal distribution. The following arguments of the list elements contain the characteritiscs of the prior distribution chosen: for "unif", two numbers must be given: the minimum and maximum values of the uniform distribution; for "normal", two numbers must be given: the mean and standard deviation of the normal distribution. Syntax adapted from the \code{EasyABC} package.}

\item{nsim}{Number of simulations}

\item{tol}{Required proportion of accepted simulations.}

\item{ncore}{Number of cores for parallel processing. Default is 1.}
}
\description{
Estimates model parameter via Approximate Bayesian Computation.
}
