% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cFreqData.R
\name{cFreqData}
\alias{cFreqData}
\title{Creates a cFreqData class object}
\usage{
cFreqData(x, timestamp, duration = 1)
}
\arguments{
\item{x}{A matrix storing the frequency of each cultural variant (columns) at each sampling phase (rows).}

\item{timestamp}{A vector containing the timestamp (i.e. last timestep) of each sampling phase.}

\item{duration}{A vector containing the durations of the sampling phase (except the first phase) calculated in number of timesteps. If a single value is supplied the durations will be assumed to be the same for all phases.}
}
\value{
An object of class cFreqData.
}
\description{
Creates a \code{cFreqData} class object from user supplied time frequency data.
}
\examples{
dt=as.matrix(data.frame(var1=c(0,10,15),var2=c(20,18,0),var3=c(0,0,4),var4=c(32,20,18)))
x=cFreqData(dt,timestamp = c(1,6,11))
}
