% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissionModels.R
\name{conformistBias}
\alias{conformistBias}
\title{Conformist-Biased Transmission Model}
\usage{
conformistBias(x, k, mu, c)
}
\arguments{
\item{x}{Relative or absolute frequency of cultural variants.}

\item{k}{Number of initial cultural variants with the last slot for mutation events (see details below).}

\item{mu}{Innovation rate.}

\item{c}{Probability of an inividual engaging into a conformist biased transmission (i.e. the selection of the most common cultural variant).}
}
\description{
Conformist-Biased Transmission Model
}
\details{
. The function estimates the probability of  each variant being selected in the tranmission process. The last slot of the user supplied vector \code{x} is designated to the outcome of innovation events.
}
