% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmissionModels.R
\name{frequencyBias}
\alias{frequencyBias}
\title{Frequency Biased Transmission Model}
\usage{
frequencyBias(x, mu, b)
}
\arguments{
\item{x}{Relative or absolute frequency of cultural variants.}

\item{mu}{Innovation rate.}

\item{b}{Transmission bias prameter. When b>0 the transmission has conformist bias, when b<0 the transmission has anticonformist bias, and when b=0 the transmission is unbiased.}
}
\description{
Single parameter frequency biased transmission model
}
\details{
. The function estimates the probability of  each variant being selected in the tranmission process. The last slot of the user supplied vector \code{x} is designated to the outcome of innovation events.
}
