% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cultural_change.R
\name{generate_cultural_change}
\alias{generate_cultural_change}
\title{Computes exected frequency of cultural variants.}
\usage{
generate_cultural_change(x, iniPop, rr, mu, params = list(b = 0),
  model = frequencyBias)
}
\arguments{
\item{x}{A cFreqData class object.}

\item{iniPop}{Raw population frequency of culturial variants}

\item{rr}{Number of individuals to be removed and added at each tranmission even. Output of \code{generate_removalReplacement}.}

\item{mu}{Innovation rate.}

\item{model}{A transmission model (see details below)}

\item{param}{A list containing parameters values to be forwarded to the transmission model.}
}
\description{
This functions computes expected frequencies of extant cultural variants given a user defined transmission model and observed data.
}
\details{
Transmission model can be either the one provided by \code{cTransmission} or custom generated. Custom generated model should be written as a function which outputs the probability of selecting each variant. The function would require at least the arguments x (the raw frequencies of each variant), k (the number of variants), and mu (the mutation rate), plus additional parameters.
}
