% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initialPop.R
\name{generate_initialPop}
\alias{generate_initialPop}
\title{Estimates the population frequencies of cultural variants.}
\usage{
generate_initialPop(x, nsim = 1, alpha = 1)
}
\arguments{
\item{x}{A vector of raw frequencies}

\item{nsim}{Number of repetitions. Default is 1.}

\item{alpha}{Dirichlet \eqn{\alpha} parameter. Default is 1.}
}
\value{
A matrix with relative frequencies of each variant (column) across \code{nsim} simulations (rows).
}
\description{
The function estimates the population relative frequencies of cultural variants given observed raw frequencies and alpha priors.
}
\keyword{internal}
