% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_popSize.R
\name{generate_popSize}
\alias{generate_popSize}
\title{Estimate Population Sizes}
\usage{
generate_popSize(x, sMean, sVariance)
}
\arguments{
\item{x}{a cFreqData class object}

\item{sMean}{mean of the ratio \eqn{s} between sample size and population size.}

\item{sVariance}{variance of the ratio \eqn{s} between sample size and population size.}
}
\value{
A vector containing the population size for each time-step.
}
\description{
Estimates population sizes from observed sample sizes given the sample fraction and interpolates across timesteps.
}
\details{
The function estimates population sizes for each timestep (from the first time-step to the start of the final time-step )
}
