% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_removalReplacement.R
\name{generate_removalReplacement}
\alias{generate_removalReplacement}
\title{Computes the number of samples to be removed at each transition.}
\usage{
generate_removalReplacement(x, N, r)
}
\arguments{
\item{N}{vector of population sizes (user defined or created using \code{generate_N})}

\item{r}{population replacement rate}

\item{cFreqData}{class object}
}
\value{
A data.frame containing the number of variants to be removed (u) and added (v) to the population at each time step.
}
\description{
Computes the number of samples to be removed at each transition.
}
