% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.pcheck.R
\name{plot.pcheck}
\alias{plot.pcheck}
\title{Plot posterior predicted checks}
\usage{
\method{plot}{pcheck}(x, var.names = NULL, phase.names = NULL,
  index = 1, hpdi = 0.95)
}
\arguments{
\item{x}{A \code{pcheck} class object obtained using \code{link{predCheck}}.}

\item{var.names}{Variant names supplied with the same order as the observed data.}

\item{phase.names}{Names of the sampling phases}

\item{index}{Integer value indicating which sampling phase to display counting from the second observed sampling phase.}

\item{hpdi}{Higher posterior interval size.}
}
\description{
Compares observed variants frequency to the expected variants frequency of the inferred transmission model.
}
\seealso{
\code{\link{predCheck}}
}
