% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPost.R
\name{plotPost}
\alias{plotPost}
\title{Plot posterior predicted checks}
\usage{
plotPost(x, pnames)
}
\arguments{
\item{x}{a matrix containing the posterior samples.}

\item{pnames}{a vector containing parameter names}
}
\description{
Plots pairwise joint and marginal distributions of ABC posterior samples.
}
