% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predCheck.R
\name{predCheck}
\alias{predCheck}
\title{Posterior predictive check for cultural transmission model.}
\usage{
predCheck(x, sim.model, posterior, nsim = NULL)
}
\arguments{
\item{x}{A cFreqData class object.}

\item{sim.model}{A simulation model created using the \code{genSim} function.}

\item{posterior}{Posterior samples. Can be extracted from the output of \code{EasyABC} functions such as \code{ABC_rejection} and \code{ABC_sequential}.}

\item{nsim}{Number of simulations for posterior predictive check.}
}
\value{
An object of class pcheck.
}
\description{
Creates a \code{pcheck} class object containing expected variant frequencies based on posterior parameter values.
}
\seealso{
\code{\link{genSim}}; \code{\link{cFreqData}}
}
