% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{rdirichlet}
\alias{rdirichlet}
\title{Generate a random samples from the Dirichlet distribution}
\usage{
rdirichlet(nsim, alpha)
}
\arguments{
\item{nsim}{Number of random vectors to generate}

\item{alpha}{Vector containing shape parameters.}
}
\value{
A matrix containing randoms samples from the Dirichlet distribution, where each row is a one simulated set.
}
\description{
Generate a random samples from the Dirichlet distribution
}
